+++ Resources/createSelfExtracting.zsh
@@ -6,14 +6,15 @@
 
 # Updated by: Dan K. Snelson
 # For Microsoft 365 Reset
-# Date: 13-Apr-2026
+# Date: 20-May-2026
 
 # Script for creating self extracting base64 encoded files.
 
 # usage: file_to_self_extracting_script <file_path> [target_path]
 
 SCRIPT_NAME=$(basename "$0")
-FILE_TO_ENCODE="$(cd "$(dirname "$0")"/.. && pwd)/Microsoft-365-Reset.zsh"
+SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
+FILE_TO_ENCODE="$(cd "${SCRIPT_DIR}"/.. && pwd)/Microsoft-365-Reset.zsh"
 TARGET_PATH="/var/tmp/Microsoft-365-Reset.zsh"
 
 datestamp=$( date '+%Y-%m-%d-%H%M%S' )
@@ -21,9 +22,10 @@ datestamp=$( date '+%Y-%m-%d-%H%M%S' )
 file_to_self_extracting_script() {
     base64_string=$(base64 -i "$1")
     filename=$(basename "$1")
-    target_path=${2}
+    local target_path="${2:-${TARGET_PATH}}"
+    output_script="${SCRIPT_DIR}/${filename}_self-extracting-${datestamp}.sh"
 
-    cat <<EOF > "${filename}_self-extracting-${datestamp}.sh"
+    cat <<EOF > "${output_script}"
 #!/bin/sh
 base64_string='$base64_string'
 echo "\$base64_string" | base64 -d > "${target_path}"
@@ -31,7 +33,7 @@ echo "File '${target_path}' has been created."
 chmod u+x "${target_path}"
 zsh "${target_path}"
 EOF
-    echo "Self-extracting script '${filename}_self-extracting-${datestamp}.sh' created."
+    echo "Self-extracting script '${output_script}' created."
 }
 
 printUsage() {